fix: tests

+++ via_verifier/lib/via_musig2/examples/coordinator.rs
@@ -451,7 +451,7 @@ async fn create_signing_session(state: &AppState) -> anyhow::Result<SigningSessi
                 Arc::new(WithdrawalFeeStrategy::new()),
                 None,
                 None,
-                MAX_STANDARD_TX_WEIGHT,
+                MAX_STANDARD_TX_WEIGHT as u64,
             )
             .await?[0]
             .clone()
