+++ via_verifier/lib/via_musig2/src/test/charge_user_fee.rs
@@ -172,7 +172,7 @@ mod tests {
                 Arc::new(WithdrawalFeeStrategy::new()),
                 None,
                 None,
-                MAX_STANDARD_TX_WEIGHT,
+                MAX_STANDARD_TX_WEIGHT as u64,
             )
             .await?[0]
             .clone();
